Divice train, serialize into train, compression, serialize
